Typo fix: "blue" -> "blur" :D & Formatting fix (#6)
I imagine you were "seeing blue" by the time you finished writing this so the typo is understandable. :D 

Also added two trailing spaces to second to last line of link list so the following new line is preserved which puts the last link on a new line as is the case with the other links.